refactor: DaoTest에서 Section 생성을 TestHelper에서 하게 한다
