+++ src/test/java/subway/domain/SectionTest.java
@@ -7,7 +7,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import subway.TestHelper;
+import subway.DomainFixture;
 
 @DisplayName("Section 클래스")
 class SectionTest {
@@ -87,8 +87,8 @@ class SectionTest {
             Station middleStation = new Station(1L, "middleStation");
             Station downStation = new Station(2L, "downStation");
 
-            Section section = TestHelper.Section.buildWithStations(upStation, middleStation);
-            Section downSection = TestHelper.Section.buildWithStations(middleStation, downStation);
+            Section section = DomainFixture.Section.buildWithStations(upStation, middleStation);
+            Section downSection = DomainFixture.Section.buildWithStations(middleStation, downStation);
 
             // when
             section.connectDownSection(downSection);
@@ -110,8 +110,8 @@ class SectionTest {
             Station differentMiddleStation = new Station(3L, "differentMiddleStation");
             Station downStation = new Station(4L, "downStation");
 
-            Section section = TestHelper.Section.buildWithStations(upStation, middleStation);
-            Section downSection = TestHelper.Section.buildWithStations(differentMiddleStation, downStation);
+            Section section = DomainFixture.Section.buildWithStations(upStation, middleStation);
+            Section downSection = DomainFixture.Section.buildWithStations(differentMiddleStation, downStation);
 
             // when
             Exception exception = catchException(() -> section.connectDownSection(downSection));
@@ -127,7 +127,7 @@ class SectionTest {
             Station upStation = new Station(2L, "upStation");
             Station middleStation = new Station(1L, "middleStation");
 
-            Section section = TestHelper.Section.buildWithStations(upStation, middleStation);
+            Section section = DomainFixture.Section.buildWithStations(upStation, middleStation);
 
             Section downSection = null;
 
@@ -151,8 +151,8 @@ class SectionTest {
             Station middleStation = new Station(1L, "middleStation");
             Station downStation = new Station(2L, "downStation");
 
-            Section section = TestHelper.Section.buildWithStations(upStation, middleStation);
-            Section downSection = TestHelper.Section.buildWithStations(middleStation, downStation);
+            Section section = DomainFixture.Section.buildWithStations(upStation, middleStation);
+            Section downSection = DomainFixture.Section.buildWithStations(middleStation, downStation);
 
             section.connectDownSection(downSection);
 
@@ -176,8 +176,8 @@ class SectionTest {
             Station middleStation = new Station(1L, "middleStation");
             Station downStation = new Station(2L, "downStation");
 
-            Section upSection = TestHelper.Section.buildWithStations(upStation, middleStation);
-            Section downSection = TestHelper.Section.buildWithStations(middleStation, downStation);
+            Section upSection = DomainFixture.Section.buildWithStations(upStation, middleStation);
+            Section downSection = DomainFixture.Section.buildWithStations(middleStation, downStation);
 
             upSection.connectDownSection(downSection);
 
@@ -196,7 +196,7 @@ class SectionTest {
             Station upStation = new Station(1L, "upStation");
             Station middleStation = new Station(1L, "middleStation");
 
-            Section section = TestHelper.Section.buildWithStations(upStation, middleStation);
+            Section section = DomainFixture.Section.buildWithStations(upStation, middleStation);
 
             // when
             Exception exception = catchException(section::disconnectDownSection);
